Added functionality for faking the necessary data in a DecodedDisk to make the visualiser work. Blank images can now be created in memory.

+++ lib/data/sector.cc
@@ -20,7 +20,7 @@ std::string Sector::statusToString(Status status)
         case Status::CONFLICT:
             return "conflicting data";
         default:
-            return fmt::format("unknown error {}", status);
+            return fmt::format("unknown error {}", (int)status);
     }
 }
